Middleware
MIDDLEWARE CONNECTION BACKEND

The file is required in order to connect to the API. For any fetches for the frontend, the backend route starts default from '/api'. I have tested it with a small backend demo to just send a small JSON package to the frontend from a backend route using useState and useEffect. This commit will be straight to master as it's required if you want to test work related to backend.